refactor: moved pages to src folder
